+++ CHANGELOG.md
@@ -9,6 +9,8 @@ Commit messages follow [Conventional Commits](https://www.conventionalcommits.or
 
 ## [Unreleased]
 
+## [1.4.4] - 2026-08-18
+
 ### Added
 
 - Reader can suggest adding a detected website RSS feed